print_meminfo accepts wchar tags
