+++ src/madness/world/worldmem.cc
@@ -38,18 +38,6 @@
 #include <climits>
 #include <iostream>
 #include <iomanip>
-#include <fstream>
-#include <sstream>
-
-#if defined(HAVE_IBMBGQ)
-#include <spi/include/kernel/memory.h>
-#elif defined(ON_A_MAC)
-#include <malloc/malloc.h>
-#elif defined(X86_32)
-#include <malloc.h>
-#elif defined(X86_64)
-#include <sys/sysinfo.h>
-#endif
 
 /*
 
@@ -119,75 +107,6 @@ namespace madness {
         max_num_frags = 0;
         cur_num_bytes = 0;
         max_num_bytes = 0;
-    }
-
-    void print_meminfo(int id, const std::string& tag,
-                       const std::string prefix) {
-      using namespace std;
-#if defined(WORLD_MEM_PROFILE_ENABLE)
-      ofstream memoryfile;
-      ostringstream filename;
-
-      filename << prefix << "." << id;
-
-      memoryfile.open(filename.str().c_str(), ios::out | ios::app);
-      memoryfile << tag << endl;
-
-      const double fac = 1024.0 * 1024.0; /* Convert from bytes to megabytes */
-#if defined(HAVE_IBMBGQ)
-      uint64_t shared, persist, heapavail, stackavail, stack, heap, guard, mmap;
-
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_SHARED, &shared);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_PERSIST, &persist);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_HEAPAVAIL, &heapavail);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_STACKAVAIL, &stackavail);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_STACK, &stack);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_HEAP, &heap);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_GUARD, &guard);
-      Kernel_GetMemorySize(KERNEL_MEMSIZE_MMAP, &mmap);
-
-      memoryfile << "Allocated heap (MB): " << (heap / fac)
-                 << ", avail. heap: " << (heapavail / fac) << endl;
-      memoryfile << "Allocated stack (MB): " << (stack / fac)
-                 << ", avail. stack: " << (stackavail / fac) << endl;
-      memoryfile << "Memory: shared: " << (shared / fac)
-                 << ", persist: " << (persist / fac)
-                 << ", guard: " << (guard / fac) << ", mmap: " << (mmap / fac)
-                 << endl;
-#elif defined(ON_A_MAC)
-      /* Mac OS X specific hack - un-tested post Snow Leopard */
-      struct malloc_statistics_t mi; /* structure in bytes */
-
-      malloc_zone_statistics(nullptr, &mi);
-
-      memoryfile << "Allocate heap (MB): " << (mi.size_in_use / fac) << endl;
-#elif defined(X86_32)
-      struct mallinfo mi; /* structure in bytes */
-
-      mi = mallinfo();
-
-      memoryfile << "Non-mmap (MB): " << (mi.arena / fac) << endl;
-      memoryfile << "Mmap (MB): " << (mi.hblkhd / fac) << endl;
-      memoryfile << "Total malloc chunks (MB): " << (mi.uordblks / fac) << endl;
-#elif defined(X86_64)
-      /* Better than nothing, mallinfo unreliable on 64-bit machine due to
-         use of int in mallinfo data structure. Requires Linux
-         kernel. Inaccurate if other processes besides MADNESS are
-         running. Memory differences appear to be reliable. */
-      struct sysinfo si; /* structure in bytes */
-
-      sysinfo(&si);
-
-      memoryfile << "Total RAM (MB): " << (si.totalram / fac) << endl;
-      memoryfile << "Free RAM (MB): " << (si.freeram / fac) << endl;
-      memoryfile << "Buffer (MB): " << (si.bufferram / fac) << endl;
-      memoryfile << "RAM in use (MB): "
-                 << ((si.totalram - si.freeram + si.bufferram) / fac) << endl;
-#endif  // platform specific
-      memoryfile.close();
-#else   // WORLD_MEM_PROFILE_ENABLE
-      return;
-#endif  // WORLD_MEM_PROFILE_ENABLE
     }
 
 }  // namespace madness
